Refactor Actor and Transform; Enhance Vector2 operations
Refactored the `Actor` class by moving its implementation to
dedicated files (`Actor.h`, `Actor.cpp`) and introduced a new
`Tranform` struct in `Tranform.h` to encapsulate position,
rotation, and scale. Updated `Game.cpp` to use the new
modularized `Actor` and `Tranform` implementations.

Enhanced the `Vector2` class with additional operators,
utility methods (e.g., `Normalize`, `Dot`, `Rotate`), and
bounds checking for safer and more versatile vector
operations.

Fixed a bug in `Random.h` by casting `RAND_MAX` to `float`
for correct random float generation. Added private members
to the `Renderer` class for better encapsulation.

Updated project files (`Engine.vcxproj`, `Engine.vcxproj.filters`)
to include the new `Actor` and `Tranform` files, improving
project organization and maintainability.

+++ Engine/Random.h
@@ -31,7 +31,7 @@ namespace nu {
 	
 	inline float RandomFloat() {
 		// rand() = 0 <- RAND_MAX
-		return rand() / RAND_MAX;
+		return rand() / (float)RAND_MAX;
 	}
 	inline float RandomFloat(float max) {
 		// 0 <-> 1 * max
